Use `go build` to validate generated code builds in makefile
