+++ internal/pkg/handlers/types.go
@@ -4,17 +4,6 @@ import (
 	"context"
 )
 
-// Register tracks all the current handlers
-// add new handlers to the array to augment the
-// processing of items in the
-var Register = []Expander{
-	&ResourceGroupResourceExpander{},
-	&SubscriptionExpander{},
-	&ActionExpander{},
-	&SwaggerResourceExpander{},
-	&DeploymentsExpander{},
-}
-
 // Expander is used to open/expand items in the left list panel
 // a single item can be expanded by 1 or more expanders
 // each Expander provides two methods.
